requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
PyYAML==6.0
scipy==1.7.1
xarray==0.19.0
